Combined initialisation of database to a single runner class.
No need to specify bean order.
Initialise Genre first, as it is largely unchanged and Film depends on it.
Initialised Star next, as it is less likely to change and Film depends on it.
Initialise films last to be able to make use of Genre and Star in future development.